+++ packages/react-essentials-commands/src/_bin/middlewares/package.middleware.ts
@@ -1,7 +1,6 @@
 import semver from "semver";
 
 import {
-  args,
   execute,
   files,
   getPackageJSON,
@@ -54,11 +53,13 @@ export default async function packageMiddleware(
 async function installDependencies(context: Context): Promise<void> {
   if (context.command !== "regenerate") return;
 
-  const fileArgs = args.get("file");
-  if (!fileArgs.length || fileArgs.includes("package.json"))
+  if (
+    !context.filesToRegenerate.length ||
+    context.filesToRegenerate.includes("package.json")
+  )
     await execute("npm i --ignore-scripts --no-audit --no-fund", false);
 
-  await files.removeFile(".npmignore");
+  if (!context.filesToRegenerate.length) await files.removeFile(".npmignore");
 }
 
 async function getTemplate(context: Context): Promise<Record<string, any>> {
